refactor: docker.gpgが存在しない場合のみdownloadするよう修正

+++ core.sh
@@ -235,7 +235,9 @@ setup_real_ubuntu() {
 
   apt-get install -y ca-certificates gnupg
   install -m 0755 -d /etc/apt/keyrings
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+  if [ ! -f "/etc/apt/keyrings/docker.gpg" ]; then
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+  fi
   chmod a+r /etc/apt/keyrings/docker.gpg
   echo \
     "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
